Fetch from stackoverflow API
Fetches lists of questions from SO. Had a look at the PHP room's code
but didn't really like it so I mostly ignored it. My code does a lot of
looping and I tried to tone it down but it's still a bit messy in the
name of having a clear way of seeing which questions are included in
what categories and keeping the number of SO API calls down to one.

I haven't styled it, I hate CSS.

There's basically no error handling.